ml_pipelines/requirements.txt
pandas
sqlalchemy
psycopg2-binary
scikit-learn
matplotlib
seaborn
jupyter
numpy
joblib
python-dotenv
